Update 81-IDevID_and_IAK_support.md
I made some minor edits and submitted a PR. Note that the reference for IDevID is old. The spec is published, show we should refer to the published spec. 

This made no sense to me, so I deleted it in the PR: “At the configuration files it should be possible to enable the IAK and IDevID validation, as specify if the steps 1 and 2 are required.” The paragraph without the sentence makes sense to me. If there is something that needs to be said, please rephrase the above sentence.

See also addition comments in my email. I wasn't sure how to fix all the issues that I raised.